+++ src/textual/css/model.py
@@ -20,6 +20,7 @@ class SelectorType(Enum):
     TYPE = 2
     CLASS = 3
     ID = 4
+    NESTED = 5
 
 
 class CombinatorType(Enum):
@@ -175,8 +176,7 @@ class RuleSet:
             elif selector.combinator == CombinatorType.CHILD:
                 tokens.append(" > ")
             tokens.append(selector.css)
-            for pseudo_class in selector.pseudo_classes:
-                tokens.append(f":{pseudo_class}")
+
         return "".join(tokens).strip()
 
     @property