+++ tests/snapshot_tests/test_snapshots.py
@@ -967,3 +967,10 @@ def test_zero_scrollbar_size(snap_compare):
     """Regression test for missing content with 0 sized scrollbars"""
     # https://github.com/Textualize/textual/issues/3886
     assert snap_compare(SNAPSHOT_APPS_DIR / "zero_scrollbar_size.py")
+
+
+def test_tree_clearing_and_expansion(snap_compare):
+    """Test the Tree.root.is_expanded state after a Tree.clear"""
+    # https://github.com/Textualize/textual/issues/3557
+    assert snap_compare(SNAPSHOT_APPS_DIR / "tree_clearing.py")
+